+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Note that not all internal changes are implemented yet to allow for a migration 
 - **Fix**: Register `MetroDiagnostics` in FIR.
 - **Fix**: Use correct severity when reporting warnings to `MessageCollector` from newer IR diagnostics factories.
 - **Fix**: When transforming FIR override statuses, check all supertypes and not just immediate supertype.
+- **Fix**: Carry qualifiers over from Dagger inject constructors when interoping with dagger factories.
 - Dagger runtime interop is enabled, do not run status transformation on `@Provides` declarations in dagger modules.
 
 0.6.6